requirements.txt
numpy
scipy
matplotlib
torch
cvxpy
qpth
roboticstoolbox-python

-e .

